WIP 3 errors left, added a new file & namespaces
